rename collapsible navigation drawer component to navigation drawer
