+++ src/app/modules/data-processing/data-processing-details/data-processing-details.component.html
@@ -19,7 +19,7 @@
     </div>
   </div>
   <div class="content">
-    <app-collapsible-navigation-drawer [items]="navigationItems"> </app-collapsible-navigation-drawer>
+    <app-navigation-drawer [items]="navigationItems"> </app-navigation-drawer>
 
     <router-outlet></router-outlet>
   </div>
