+++ src/app/modules/it-systems/it-system-catalog/it-system-catalog-details/it-system-catalog-details.component.html
@@ -62,7 +62,7 @@
   </app-details-header>
 
   <div class="content">
-    <app-collapsible-navigation-drawer [items]="navigationItems"> </app-collapsible-navigation-drawer>
+    <app-navigation-drawer [items]="navigationItems"> </app-navigation-drawer>
 
     <router-outlet></router-outlet>
   </div>
