+++ src/app/modules/it-systems/it-system-catalog/it-system-catalog-details/it-system-catalog-details.component.ts
@@ -27,7 +27,7 @@ import {
   selectItSystemUuid,
 } from 'src/app/store/it-system/selectors';
 import { ITSystemCatalogDetailsComponentStore } from './it-system-catalog-details.component-store';
-import { NavigationDrawerItem } from 'src/app/shared/components/collapsible-navigation-drawer/collapsible-navigation-drawer.component';
+import { NavigationDrawerItem } from 'src/app/shared/components/navigation-drawer/navigation-drawer.component';
 
 @Component({
   templateUrl: './it-system-catalog-details.component.html',
