+++ src/app/modules/it-systems/it-system-interfaces/it-system-interfaces-details/it-system-interfaces-details.component.html
@@ -41,7 +41,7 @@
       </ng-container>
     </app-details-header>
   <div class="content">
-    <app-collapsible-navigation-drawer [items]="navigationItems"> </app-collapsible-navigation-drawer>
+    <app-navigation-drawer [items]="navigationItems"> </app-navigation-drawer>
 
     <router-outlet></router-outlet>
   </div>
