+++ src/app/modules/it-systems/it-system-usages/it-system-usage-details/it-system-usage-details.component.html
@@ -13,7 +13,7 @@
     >
   </div>
   <div class="content">
-    <app-collapsible-navigation-drawer [items]="navigationItems"> </app-collapsible-navigation-drawer>
+    <app-navigation-drawer [items]="navigationItems"> </app-navigation-drawer>
 
     <router-outlet></router-outlet>
   </div>
