+++ src/app/modules/it-systems/it-system-usages/it-system-usage-details/it-system-usage-details.component.ts
@@ -21,7 +21,7 @@ import {
 import { ITSystemActions } from 'src/app/store/it-system/actions';
 import { selectOrganizationName } from 'src/app/store/user-store/selectors';
 import { ITSystemUsageRemoveComponent } from './it-system-usage-remove/it-system-usage-remove.component';
-import { NavigationDrawerItem } from 'src/app/shared/components/collapsible-navigation-drawer/collapsible-navigation-drawer.component';
+import { NavigationDrawerItem } from 'src/app/shared/components/navigation-drawer/navigation-drawer.component';
 
 @Component({
   templateUrl: 'it-system-usage-details.component.html',
