+++ src/app/modules/local-admin/local-admin.component.html
@@ -1,6 +1,6 @@
 <h4 i18n class="title-margin">Lokal admin</h4>
 <div class="content">
-  <app-collapsible-navigation-drawer [items]="items"> </app-collapsible-navigation-drawer>
+  <app-navigation-drawer [items]="items"> </app-navigation-drawer>
 
   <router-outlet></router-outlet>
 </div>
