+++ src/app/modules/local-admin/local-admin.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { NavigationDrawerItem } from 'src/app/shared/components/collapsible-navigation-drawer/collapsible-navigation-drawer.component';
+import { NavigationDrawerItem } from 'src/app/shared/components/navigation-drawer/navigation-drawer.component';
 import { AppPath } from 'src/app/shared/enums/app-path';
 
 @Component({
