+++ src/app/shared/components/components.module.ts
@@ -145,7 +145,7 @@ import { RoleTableContainerComponent } from './role-table/role-table-container/r
 import { SlideToggleComponent } from './slide-toggle/slide-toggle.component';
 import { SnackbarActionsComponent } from './snackbar-actions/snackbar-actions.component';
 import { VerticalContentGridSectionMarginLeftComponent } from './vertical-content-grid-section-margin-left/vertical-content-grid-section-margin-left.component';
-import { CollapsibleNavigationDrawerComponent } from './collapsible-navigation-drawer/collapsible-navigation-drawer.component';
+import { NavigationDrawerComponent } from './navigation-drawer/navigation-drawer.component';
 import { IconComponent } from './icon/icon.component';
 
 export function scrollFactory(overlay: Overlay): () => RepositionScrollStrategy {
@@ -261,7 +261,7 @@ export function scrollFactory(overlay: Overlay): () => RepositionScrollStrategy 
     AccordionHeaderComponent,
     OptionTypeTableComponent,
     EditOptionTypeDialogComponent,
-    CollapsibleNavigationDrawerComponent,
+    NavigationDrawerComponent,
     IconComponent,
   ],
   imports: [
@@ -392,7 +392,7 @@ export function scrollFactory(overlay: Overlay): () => RepositionScrollStrategy 
     SnackbarActionsComponent,
     AccordionHeaderComponent,
     OptionTypeTableComponent,
-    CollapsibleNavigationDrawerComponent,
+    NavigationDrawerComponent,
     IconComponent,
   ],
   providers: [
