+++ src/app/shared/components/navigation-drawer/navigation-drawer.component.scss
